requirements.txt
Jinja2
PyICU
sanic
